pyproject.toml
[project]
name = "fender"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "passlib>=1.7.4",
    "pydantic-settings>=2.9.1",
    "pyjwt>=2.10.1",
    "pynamodb>=6.0.2",
    "pytest>=8.3.5",
]
